core refactorings
