+++ docs/content/docs/concepts/code-structure.md
@@ -17,16 +17,17 @@ File Dependencies:
 
 graph BT
     platform/index --> foundation/index
+    core/domains --> core/classes
     core/objectid --> core/classes
     core/rpc --> foundation/index
-    core/space --> core/classes
-    core/text --> core/classes & core/tx & core/objectid & core/model
-    core/title --> core/classes & core/tx & core/objectid & core/model
-    core/vdoc --> core/classes & core/tx & core/objectid & core/model
-    core/tx --> core/model & core/classes
-    core/model --> core/classes & core/tx
+    core/tx --> core/classes
+    core/model --> core/classes & core/tx & core/objectid
     core/textmodel
     core/utils --> core/classes & core/tx
+    core/indices/model --> core/classes & core/tx & core/model
+    core/indices/vdoc --> core/tx & core/model
+    core/indices/text --> core/classes & core/tx & core/objectid & core/model & core/domains
+    core/indices/title --> core/classes & core/tx & core/objectid & core/model & core/domains
 
 {{< /mermaid >}}
