+++ packages/core/src/index.ts
@@ -20,11 +20,8 @@ import type { Obj, Emb, Doc, Classifier, Attribute, Mixin, Type, ArrayOf, RefTo,
 import model from '@anticrm/model'
 import { Class, Ref } from './classes'
 import { AnyPlugin, identify } from '@anticrm/platform'
-import { Backlinks } from './text'
-import { Title } from './title'
+import { Backlinks, Application, VDoc, Space, SpaceUser, Title } from './domains'
 import { CreateTx, DeleteTx, PushTx, UpdateTx, Tx } from './tx'
-import { Application, VDoc } from './vdoc'
-import { Space, SpaceUser } from './space'
 import { mergeWith } from 'lodash'
 
 type Namespace = Record<string, Record<string, any>>
@@ -62,11 +59,8 @@ export default core
 export * from './rpc'
 export * from './classes'
 export * from './model'
-export * from './text'
 export * from './textmodel'
 export * from './tx'
-export * from './title'
-export * from './space'
-export * from './vdoc'
+export * from './domains'
 export * from './objectid'
 export * from './utils'
