+++ plugins/task/src/index.ts
@@ -1,5 +1,5 @@
 //
-// Copyright © 2020 Anticrm Platform Contributors.
+// Copyright © 2020, 2021 Anticrm Platform Contributors.
 //
 // Licensed under the Eclipse Public License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License. You may
@@ -30,7 +30,7 @@ export interface TaskService extends Service {
 export default plugin('task' as Plugin<TaskService>, {}, {
   icon: {
     Task: '' as Asset,
-    ArrowDown: '' as Asset,
+    ArrowDown: '' as Asset
   },
   class: {
     Task: '' as Ref<Class<Task>>
